[Tsavorite] Use int consistently for page numbers
Page numbers were an inconsistent mix of `long` and `int` throughout Tsavorite.
`PageOffset.Page` is an `int` and `TailPageOffset` is the sole allocation path
(`TryAllocate` -> `Interlocked.Add(ref TailPageOffset.PageAndOffset, numSlots)`,
`HandlePageOverflow` -> `localTailPageOffset.Page++`), so no allocator derived
from `AllocatorBase` can ever produce a page number above `int.MaxValue`. This
narrows page numbers to `int` everywhere so the type matches that invariant and
the repeated `(int)` narrowing casts at page-index sites disappear.

### Description of Change

Foundational signatures were narrowed first, then the compiler was used as an
exhaustive oracle for the downstream sites (implicit `long` -> `int` narrowing is
a compile error, so nothing can be silently missed).

**Key changes:**
- `LogAddress.GetPageOfAddress` returns `int`; `GetLogicalAddressOfStartOfPage`
  takes an `int` page and widens before shifting.
- `IAllocator` and the three allocator wrapper structs (`ObjectAllocator`,
  `SpanByteAllocator`, `TsavoriteLogAllocator`) updated to match.
  `TsavoriteLogAllocator.GetPageOfAddress` remains deliberately unmasked - it
  must not strip bit 47.
- `AllocatorBase`: `GetPage`, `GetPageIndexForPage`,
  `GetFirstValidLogicalAddressOnPage`, `ClearPage`, `NeedToShiftAddress`,
  `IssueShiftAddress`, `EvictPageForRecovery`, `FindHeadAddressCutoffOnPage`,
  `GetFlushPageRange`, `PrepareFlushAsyncResult`, `AsyncFlushPagesForRecovery`,
  `AsyncFlushPagesForSnapshot`, `AsyncReadPagesForRecovery`,
  `AsyncReadPageFromDeviceToFrame`, `WriteAsync`, `WriteAsyncToDeviceForSnapshot`,
  `CalculatePageObjectSizes`, `LoadObjectsForRecoveryPass2` and
  `GetLowestObjectLogPositionForPage`.
- `Recovery`, `SnapshotFlushCoordination`, `ScanIteratorBase` and the three scan
  iterators, `TsavoriteLog`, `LogSizeTracker`, `PageAsyncResultTypes`.

### Key Technical Details

**New `AllocatorBase.GetFileOffsetOfPage(int page)`.** `AlignedPageSizeBytes` is
an `int` field, so once page variables became `int` the existing
`AlignedPageSizeBytes * page` device-offset expressions silently became 32-bit
multiplies and would overflow past the first 2 GB of a log file. All such sites
now route through `GetFileOffsetOfPage`, which widens before multiplying. This
was the one real hazard of the refactor; the alternative (widening the
`AlignedPageSizeBytes` field to `long`) was rejected because it obscures that the
value genuinely is bounded by 2^27.

**`devicePageOffset` / `recoveryDevicePageOffset` are page numbers, not byte
offsets.** Despite the "Offset" suffix they hold a page number (see
`Recovery.cs`: `recoveryDevicePageOffset = snapshotStartPage`, and the
`readPage - devicePageOffset` uses), so they were narrowed to `int` as well.

**`PageAsyncFlushResult<TContext>.page` is public API surface** and changes from
`long` to `int`.

**Test change.** `TsavoriteLogPageIsNotReadCacheBitMaskedTest` asserted that the
computed page exceeded `int.MaxValue`, which is now unrepresentable. Its actual
purpose - proving `TsavoriteLog` does *not* mask off bit 47 the way the
main-store path does - is preserved by raising `PageSizeBits` from 14 to 17 (the
minimum satisfying `2^47 >> bits < 2^31`, with `MemorySizeBits`/`SegmentSizeBits`
raised to 19 to stay >= the page size) and asserting the exact expected page
value instead of an `int.MaxValue` bound.

### What NOT to Do (for future agents)

- Don't assume names containing "page" are page numbers. Byte offsets, addresses,
  sizes and page *counts* (`pageStartLogicalAddress`, `offsetInPage`,
  `maxAddressOffsetOnPage`, `ScanIteratorBase.loadedPages`,
  `PageOffset.PageAndOffset`) must stay `long` and were deliberately left alone.
- Don't multiply an `int` page by an `int` byte size to get a file offset. Use
  `GetFileOffsetOfPage`.
- Don't shift an `int` page without widening first: `page << LogPageSizeBits` is a
  32-bit shift. All page shifts are `(long)`-guarded.
- Don't do bulk scripted edits with PowerShell `[IO.File]::WriteAllText` without
  re-checking the UTF-8 BOM afterwards; the default encoding drops it and
  silently rewrites the license header line.

### Out of Scope

`FreePage`'s overloading of page index vs. page number is left as-is; the
implementations apply `% BufferSize`, which is idempotent for indices, so both
call styles work. `ScanIteratorBase.currentFrame` and the frame-index parameters
also remain `long`.

### Validation

`Tsavorite.slnx` and `Garnet.slnx` build clean on net8.0 and net10.0;
`dotnet format --verify-no-changes` clean on both. Tests (net10.0, Debug):
`Tsavorite.test.recovery` 479 passed, `Tsavorite.test.hlog` 593 passed,
`Tsavorite.test` 342 passed, `test.recordops` 379, `test.session` 155,
`test.session.context` 133, `test.epoch` 15, and `Garnet.test` 1053 passed -
0 failures across all suites.

+++ libs/storage/Tsavorite/cs/src/core/Allocator/ObjectAllocator.cs
@@ -110,14 +110,14 @@ namespace Tsavorite.core
 
         /// <inheritdoc/>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public readonly void FreePage(long pageIndex) => _this.FreePage(pageIndex);
+        public readonly void FreePage(int pageIndex) => _this.FreePage(pageIndex);
 
         /// <inheritdoc/>        
         public readonly int OverflowPageCount => _this.OverflowPageCount;
 
         /// <inheritdoc/>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public readonly long GetPageOfAddress(long logicalAddress, int logPageSizeBits) => LogAddress.GetPageOfAddress(logicalAddress, logPageSizeBits);
+        public readonly int GetPageOfAddress(long logicalAddress, int logPageSizeBits) => LogAddress.GetPageOfAddress(logicalAddress, logPageSizeBits);
 
         /// <inheritdoc/>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
@@ -134,7 +134,7 @@ namespace Tsavorite.core
         public readonly ObjectIdMap TransientObjectIdMap => _this.transientObjectIdMap;
 
         /// <inheritdoc/>
-        public readonly ObjectIdMap GetPageObjectIdMap(long pageNumber) => _this.objectPages[_this.GetPageIndexForPage(pageNumber)].objectIdMap;
+        public readonly ObjectIdMap GetPageObjectIdMap(int pageNumber) => _this.objectPages[_this.GetPageIndexForPage(pageNumber)].objectIdMap;
 
         /// <inheritdoc/>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
